Move test server to a separately importable module
