+++ tests/conftest.py
@@ -1,8 +1,5 @@
-import asyncio
 import json
 import os
-import threading
-import time
 import typing
 
 import pytest
@@ -15,10 +12,9 @@ from cryptography.hazmat.primitives.serialization import (
     load_pem_private_key,
 )
 from uvicorn.config import Config
-from uvicorn.main import Server
 
-from httpx import URL
-from tests.concurrency import sleep
+from .concurrency import sleep
+from .server import Server, serve_in_thread
 
 ENVIRONMENT_VARIABLES = {
     "SSL_CERT_FILE",
@@ -224,74 +220,18 @@ def cert_encrypted_private_key_file(localhost_cert):
         yield tmp
 
 
-class TestServer(Server):
-    @property
-    def url(self) -> URL:
-        protocol = "https" if self.config.is_ssl else "http"
-        return URL(f"{protocol}://{self.config.host}:{self.config.port}/")
-
-    def install_signal_handlers(self) -> None:
-        # Disable the default installation of handlers for signals such as SIGTERM,
-        # because it can only be done in the main thread.
-        pass
-
-    async def serve(self, sockets=None):
-        self.restart_requested = asyncio.Event()
-
-        loop = asyncio.get_event_loop()
-        tasks = {
-            loop.create_task(super().serve(sockets=sockets)),
-            loop.create_task(self.watch_restarts()),
-        }
-        await asyncio.wait(tasks)
-
-    async def restart(self) -> None:
-        # This coroutine may be called from a different thread than the one the
-        # server is running on, and from an async environment that's not asyncio.
-        # For this reason, we use an event to coordinate with the server
-        # instead of calling shutdown()/startup() directly, and should not make
-        # any asyncio-specific operations.
-        self.started = False
-        self.restart_requested.set()
-        while not self.started:
-            await sleep(0.2)
-
-    async def watch_restarts(self):
-        while True:
-            if self.should_exit:
-                return
-
-            try:
-                await asyncio.wait_for(self.restart_requested.wait(), timeout=0.1)
-            except asyncio.TimeoutError:
-                continue
-
-            self.restart_requested.clear()
-            await self.shutdown()
-            await self.startup()
-
-
-def serve_in_thread(server: Server):
-    thread = threading.Thread(target=server.run)
-    thread.start()
-    try:
-        while not server.started:
-            time.sleep(1e-3)
-        yield server
-    finally:
-        server.should_exit = True
-        thread.join()
-
-
 @pytest.fixture(scope=SERVER_SCOPE)
-def server():
+def server() -> typing.Iterator[Server]:
     config = Config(app=app, lifespan="off", loop="asyncio")
-    server = TestServer(config=config)
-    yield from serve_in_thread(server)
+    server = Server(config=config)
+    with serve_in_thread(server):
+        yield server
 
 
 @pytest.fixture(scope=SERVER_SCOPE)
-def https_server(cert_pem_file, cert_private_key_file):
+def https_server(
+    cert_pem_file: str, cert_private_key_file: str
+) -> typing.Iterator[Server]:
     config = Config(
         app=app,
         lifespan="off",
@@ -301,5 +241,6 @@ def https_server(cert_pem_file, cert_private_key_file):
         port=8001,
         loop="asyncio",
     )
-    server = TestServer(config=config)
-    yield from serve_in_thread(server)
+    server = Server(config=config)
+    with serve_in_thread(server):
+        yield server
